Refactor: MOD nested class
- dao 추가시마다 PreparedStatementStrategy를 구현한 클래스가 추가되어야 한다. 이 클래스는 dao의 dml 함수만 호출하므로
- 함수에 익명클래스를 생성해서 복잡성을 낮출 수 있다.
