--- staek-jdbc/src/main/java/com/staekframework/test/strategy/DeleteAllStrategy.java
@@ -1,14 +1,0 @@
-package com.staekframework.test.strategy;
-
-import com.staekframework.jdbc.PreparedStatementStrategy;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-public class DeleteAllStrategy implements PreparedStatementStrategy {
-    @Override
-    public PreparedStatement newStatement(Connection conn) throws SQLException {
-        return conn.prepareStatement("delete from user");
-    }
-}
